缝缝补补

+++ mvvmhabit/src/main/java/me/goldze/mvvmhabit/http/net/entity/JobHuntingEntity.java
@@ -365,7 +365,7 @@ public class JobHuntingEntity extends BaseEntity implements Serializable {
             private Object mail;
             private int age;
             private int buildNum;
-            private int cumulativeLoginTime;
+            private long cumulativeLoginTime;
             private int id;
             private String lastLogin;
             private int enable;
@@ -674,11 +674,11 @@ public class JobHuntingEntity extends BaseEntity implements Serializable {
                 this.buildNum = buildNum;
             }
 
-            public int getCumulativeLoginTime() {
+            public long getCumulativeLoginTime() {
                 return cumulativeLoginTime;
             }
 
-            public void setCumulativeLoginTime(int cumulativeLoginTime) {
+            public void setCumulativeLoginTime(long cumulativeLoginTime) {
                 this.cumulativeLoginTime = cumulativeLoginTime;
             }
 
@@ -861,7 +861,7 @@ public class JobHuntingEntity extends BaseEntity implements Serializable {
             private Object mail;
             private int age;
             private int buildNum;
-            private int cumulativeLoginTime;
+            private long cumulativeLoginTime;
             private int id;
             private String lastLogin;
             private int enable;
@@ -1170,11 +1170,11 @@ public class JobHuntingEntity extends BaseEntity implements Serializable {
                 this.buildNum = buildNum;
             }
 
-            public int getCumulativeLoginTime() {
+            public long getCumulativeLoginTime() {
                 return cumulativeLoginTime;
             }
 
-            public void setCumulativeLoginTime(int cumulativeLoginTime) {
+            public void setCumulativeLoginTime(long cumulativeLoginTime) {
                 this.cumulativeLoginTime = cumulativeLoginTime;
             }
