+++ mvvmhabit/src/main/java/me/goldze/mvvmhabit/http/net/entity/ReviewListEntity.java
@@ -261,7 +261,7 @@ public class ReviewListEntity extends BaseEntity{
             private Object mail;
             private int age;
             private int buildNum;
-            private int cumulativeLoginTime;
+            private long cumulativeLoginTime;
             private Object wxappProvince;
             private int enable;
             private Object wechatOpenId;
@@ -641,11 +641,11 @@ public class ReviewListEntity extends BaseEntity{
                 this.buildNum = buildNum;
             }
 
-            public int getCumulativeLoginTime() {
+            public long getCumulativeLoginTime() {
                 return cumulativeLoginTime;
             }
 
-            public void setCumulativeLoginTime(int cumulativeLoginTime) {
+            public void setCumulativeLoginTime(long cumulativeLoginTime) {
                 this.cumulativeLoginTime = cumulativeLoginTime;
             }
